test_live_mexc.py
#!/usr/bin/env python3
"""
🔥 LIVE MEXC DATA TEST
Test connection to MEXC API and get REAL current prices
"""

import asyncio
import aiohttp
import json
from datetime import datetime
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class LiveMexcTest:
    """Test live MEXC API connection"""
    
    def __init__(self):
        self.base_url = "https://api.mexc.com"
        self.api_key = os.getenv('MEXC_API_KEY')
        self.api_secret = os.getenv('MEXC_API_SECRET')
        
    async def get_current_prices(self):
        """Get current live prices from MEXC (no auth required)"""
        
        symbols = ['BTCUSDT', 'ETHUSDT', 'SOLUSDT', 'AVAXUSDT', 'ADAUSDT']
        
        print("🔥 GETTING LIVE MEXC PRICES...")
        print("=" * 50)
        
        async with aiohttp.ClientSession() as session:
            for symbol in symbols:
                try:
                    # Get ticker data (public API, no auth needed)
                    url = f"{self.base_url}/api/v3/ticker/24hr?symbol={symbol}"
                    
                    async with session.get(url) as response:
                        if response.status == 200:
                            data = await response.json()
                            
                            price = float(data['lastPrice'])
                            change_pct = float(data['priceChangePercent'])
                            volume = float(data['volume'])
                            
                            print(f"📈 {symbol:12} ${price:>10,.2f} ({change_pct:+6.2f}%) Vol: {volume:>15,.0f}")
                            
                        else:
                            print(f"❌ {symbol}: Error {response.status}")
                            
                except Exception as e:
                    print(f"❌ {symbol}: {e}")
        
        print("=" * 50)
        print(f"⏰ Data fetched at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    
    async def test_websocket_feed(self):
        """Test WebSocket connection for real-time data"""
        
        print("\n🌊 TESTING WEBSOCKET REAL-TIME FEED...")
        print("=" * 50)
        
        try:
            import websockets
            
            # MEXC WebSocket URL for real-time data
            ws_url = "wss://wbs.mexc.com/ws"
            
            # Subscribe to BTC price updates
            subscribe_msg = {
                "method": "SUBSCRIPTION",
                "params": ["btcusdt@ticker"]
            }
            
            async with websockets.connect(ws_url) as websocket:
                # Send subscription
                await websocket.send(json.dumps(subscribe_msg))
                print("✅ Connected to MEXC WebSocket!")
                print("📡 Subscribed to BTCUSDT real-time updates...")
                
                # Listen for a few updates
                for i in range(5):
                    try:
                        message = await asyncio.wait_for(websocket.recv(), timeout=10.0)
                        data = json.loads(message)
                        
                        if 'data' in data and 'c' in data['data']:
                            price = float(data['data']['c'])
                            print(f"🔥 LIVE BTC Price: ${price:,.2f} (Update {i+1})")
                        
                    except asyncio.TimeoutError:
                        print("⚠️ No data received (timeout)")
                        break
                    except Exception as e:
                        print(f"⚠️ WebSocket error: {e}")
                        
        except ImportError:
            print("❌ websockets library not available")
        except Exception as e:
            print(f"❌ WebSocket connection failed: {e}")
    
    async def test_api_auth(self):
        """Test API authentication (if keys are provided)"""
        
        if not self.api_key or not self.api_secret:
            print("\n⚠️ MEXC API keys not found - using public data only")
            return
        
        print(f"\n🔐 TESTING API AUTHENTICATION...")
        print("=" * 50)
        print(f"API Key: {self.api_key[:10]}...")
        
        # Test authenticated endpoint (account info)
        try:
            import hmac
            import hashlib
            import time
            
            timestamp = str(int(time.time() * 1000))
            
            # For MEXC, we'll just test if the keys are valid format
            if len(self.api_key) > 10 and len(self.api_secret) > 10:
                print("✅ API keys appear to be valid format")
                print("💡 Note: Full auth testing requires valid production keys")
            else:
                print("⚠️ API keys may be test/demo keys")
                
        except Exception as e:
            print(f"❌ Auth test failed: {e}")

async def main():
    """Main test function"""
    
    print("🚀 MEXC LIVE DATA CONNECTION TEST")
    print("🎯 Testing connection to REAL market data")
    print()
    
    tester = LiveMexcTest()
    
    # Test 1: Get current live prices (public API)
    await tester.get_current_prices()
    
    # Test 2: Test WebSocket real-time feed
    await tester.test_websocket_feed()
    
    # Test 3: Test API authentication
    await tester.test_api_auth()
    
    print("\n" + "=" * 50)
    print("🎯 COMPARISON WITH YOUR PAPER TRADING PRICES:")
    print("Paper Trading BTC: $42,549 (OLD)")
    print("Paper Trading ETH: $2,647 (OLD)")
    print("Paper Trading SOL: $104 (OLD)")
    print()
    print("👆 The live prices above show the REAL current market!")
    print("🔥 Next: Let's connect your bot to these LIVE feeds!")

if __name__ == "__main__":
    asyncio.run(main())
