requirements.txt
# Core Dependencies
aiohttp>=3.8.0
websockets>=11.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
ta>=0.10.2
ccxt>=4.0.0

# Machine Learning & AI (Optional - comment out if not needed)
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# torch>=2.0.0
# xgboost>=1.7.0
# optuna>=3.2.0

# Database & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Security & Encryption
cryptography>=41.0.0
pycryptodome>=3.18.0

# Monitoring & Alerts
psutil>=5.9.0

# Visualization & Charts
matplotlib>=3.7.0
mplfinance>=0.12.0
plotly>=5.0.0
dash>=2.15.0
seaborn>=0.12.0

# Web Interface (Optional)
# fastapi>=0.100.0
# uvicorn>=0.22.0
# jinja2>=3.1.0
websocket-client>=1.6.0

# Configuration & Utils
pydantic>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
